+++ docs/source/website_config/_specimen_web_config.py
@@ -15,6 +15,7 @@ from typing import List, TYPE_CHECKING
 # Include the following if you want to use it in CELERYBEAT_SCHEDULE
 # from celery.schedules import crontab
 
+from crate_anon.common.constants import mebibytes
 from crate_anon.crateweb.config.constants import ResearchDbInfoKeys as RDIKeys
 
 if TYPE_CHECKING:
@@ -324,7 +325,7 @@ PRIVATE_FILE_STORAGE_ROOT = "/srv/crate_filestorage"
 XSENDFILE = False
 
 # How big will we accept?
-MAX_UPLOAD_SIZE_BYTES = 10 * 1024 * 1024  # 10 Mb
+MAX_UPLOAD_SIZE_BYTES = mebibytes(10)
 
 
 # =============================================================================